refactor: extract shared CurrentSchemaVersion for RDBMS/ES schema-version providers
RdbmsSchemaMigrationStatusProvider (db/rdbms) and
ElasticsearchSchemaMigrationStatusProvider (schema-manager) each
carried their own private, nested CurrentSchemaVersion record plus an
almost byte-for-byte identical toMigrationStatus(CurrentSchemaVersion)
mapping method -- the two implementations only diverge in how they
read the persisted version (JDBC vs. a search-engine metadata
document), not in the shape of the result or how it maps to an
upgrade-readiness condition.

Extracted the shared record (+ its Kind enum and factory methods) and
the mapping logic into io.camunda.zeebe.util.migration.CurrentSchemaVersion,
alongside VersionCompatibilityCheck/SemanticVersion, which both
db/rdbms and schema-manager already depend on. No new pom.xml
dependency was needed: zeebe-util's pom.xml already has a main-scope
dependency on camunda-cluster (used today by
io.camunda.zeebe.util.micrometer.PartitionKeyNames for PartitionId),
the same module that defines MigrationConditionStatus/MigrationState,
and camunda-cluster has no dependency back on zeebe-util, so there is
no circularity.

zeebe/exporter-common was considered as a home and rejected: it is
exporter-internal machinery (audit-log transformers, entity caches,
waitstate handling) unrelated to schema-version tracking, and
db/rdbms has no reason to depend on it.

Both providers' toMigrationStatus() mapping method is removed
entirely; getMigrationStatus() now just calls
versionStore.getCurrentSchemaVersion().toMigrationStatus(). The
exhaustive per-branch mapping coverage (same version, patch/minor
upgrade, fresh database, incompatible/indeterminate versions, read
failures) moves to a single new CurrentSchemaVersionTest in
zeebe-util, tested directly against hand-built CurrentSchemaVersion
instances with no mocking needed. The two providers' own tests are
trimmed to what's still genuinely provider-specific: conditionName,
multi-tenant map-shaping, and one smoke test confirming each is wired
to the shared mapping correctly.

Pure structural change, no behavior change. Raised during review of
#60598: https://github.com/camunda/camunda/pull/60598#discussion_r3828238446

Closes #60710.
