+++ db/rdbms/src/test/java/io/camunda/db/rdbms/RdbmsSchemaMigrationStatusProviderTest.java
@@ -20,7 +20,14 @@ import java.util.Map;
 import javax.sql.DataSource;
 import org.junit.jupiter.api.Test;
 
-/** Unit tests for the RDBMS upgrade-readiness condition, one entry per physical tenant. */
+/**
+ * Unit tests for the RDBMS upgrade-readiness condition, one entry per physical tenant. The
+ * schema-version-to-{@link MigrationState} mapping itself (patch/minor upgrade, fresh database,
+ * incompatible/indeterminate versions, read failures) is shared across every schema-version-backed
+ * provider and exhaustively covered once elsewhere -- these tests only cover what's genuinely
+ * specific to this provider: multi-tenant map-shaping and one smoke test confirming it's wired to
+ * that shared mapping correctly.
+ */
 class RdbmsSchemaMigrationStatusProviderTest {
 
   @Test
@@ -62,77 +69,6 @@ class RdbmsSchemaMigrationStatusProviderTest {
     // then
     assertThat(status.state()).isEqualTo(MigrationState.MIGRATED);
     assertThat(status.detail()).contains("8.10.0");
-  }
-
-  @Test
-  void shouldReportMigrationInProgressForPatchUpgrade() {
-    // given - schema=8.9.0, app=8.9.5 (not yet migrated to the running app's exact version)
-    final var status = singleStatus(versionStore("8.9.0", "8.9.5"));
-
-    // then
-    assertThat(status.state()).isEqualTo(MigrationState.MIGRATION_IN_PROGRESS);
-  }
-
-  @Test
-  void shouldReportMigrationInProgressForMinorUpgrade() {
-    // given - schema=8.9.1, app=8.10.0
-    final var status = singleStatus(versionStore("8.9.1", "8.10.0"));
-
-    // then
-    assertThat(status.state()).isEqualTo(MigrationState.MIGRATION_IN_PROGRESS);
-  }
-
-  @Test
-  void shouldReportMigrationInProgressForFreshDatabase() {
-    // given - resolveCurrentSchemaVersion returns null (fresh DB, not yet initialized)
-    final var status = singleStatus(versionStore(null, "8.11.0"));
-
-    // then
-    assertThat(status.state()).isEqualTo(MigrationState.MIGRATION_IN_PROGRESS);
-    assertThat(status.detail()).contains("fresh database");
-  }
-
-  @Test
-  void shouldReportUnknownForIncompatibleUpgradePath() {
-    // given - schema=8.9.0, app=8.11.0 (skipped 8.10) - a real problem, not "in progress"
-    final var status = singleStatus(versionStore("8.9.0", "8.11.0"));
-
-    // then
-    assertThat(status.state()).isEqualTo(MigrationState.UNKNOWN);
-  }
-
-  @Test
-  void shouldReportUnknownForIndeterminateSchemaVersion() {
-    // given - stored schema version is not a valid semantic version
-    final var status = singleStatus(versionStore("not-a-semver", "8.10.0"));
-
-    // then
-    assertThat(status.state()).isEqualTo(MigrationState.UNKNOWN);
-  }
-
-  @Test
-  void shouldReportUnknownForUnparseableApplicationVersion() {
-    // given - app=development (not a semantic version)
-    final var status = singleStatus(versionStore("8.9.0", "development"));
-
-    // then
-    assertThat(status.state()).isEqualTo(MigrationState.UNKNOWN);
-    assertThat(status.detail()).contains("development");
-  }
-
-  @Test
-  void shouldReportUnknownWhenReadingCurrentVersionFails() throws Exception {
-    // given - getConnection() throws
-    final var dataSource = mock(DataSource.class);
-    when(dataSource.getConnection()).thenThrow(new RuntimeException("DB connection refused"));
-    final var store = new RdbmsSchemaVersionStore(dataSource, "", "8.10.0");
-
-    // when
-    final var status = singleStatus(store);
-
-    // then - a read failure must never throw; it must be reported as UNKNOWN
-    assertThat(status.state()).isEqualTo(MigrationState.UNKNOWN);
-    assertThat(status.detail()).contains("DB connection refused");
   }
 
   private static RdbmsSchemaMigrationStatusProvider provider(
