+++ schema-manager/src/test/java/io/camunda/search/schema/ElasticsearchSchemaMigrationStatusProviderTest.java
@@ -19,7 +19,14 @@ import java.util.LinkedHashMap;
 import java.util.Map;
 import org.junit.jupiter.api.Test;
 
-/** Unit tests for the Elasticsearch/OpenSearch upgrade-readiness condition. */
+/**
+ * Unit tests for the Elasticsearch/OpenSearch upgrade-readiness condition. The
+ * schema-version-to-{@link MigrationState} mapping itself (patch/minor upgrade, fresh database,
+ * incompatible/indeterminate versions, read failures) is shared across every schema-version-backed
+ * provider and exhaustively covered once elsewhere -- these tests only cover what's genuinely
+ * specific to this provider: multi-tenant map-shaping and one smoke test confirming it's wired to
+ * that shared mapping correctly.
+ */
 class ElasticsearchSchemaMigrationStatusProviderTest {
 
   @Test
@@ -61,78 +68,6 @@ class ElasticsearchSchemaMigrationStatusProviderTest {
     // then
     assertThat(status.state()).isEqualTo(MigrationState.MIGRATED);
     assertThat(status.detail()).contains("8.10.0");
-  }
-
-  @Test
-  void shouldReportMigrationInProgressForPatchUpgrade() {
-    // given - schema=8.9.0, app=8.9.5 (not yet migrated to the running app's exact version)
-    final var status = singleStatus(versionStore("8.9.0", "8.9.5"));
-
-    // then
-    assertThat(status.state()).isEqualTo(MigrationState.MIGRATION_IN_PROGRESS);
-  }
-
-  @Test
-  void shouldReportMigrationInProgressForMinorUpgrade() {
-    // given - schema=8.9.1, app=8.10.0
-    final var status = singleStatus(versionStore("8.9.1", "8.10.0"));
-
-    // then
-    assertThat(status.state()).isEqualTo(MigrationState.MIGRATION_IN_PROGRESS);
-  }
-
-  @Test
-  void shouldReportMigrationInProgressForFreshDatabase() {
-    // given - metadata index does not exist yet (fresh installation)
-    final var status = singleStatus(versionStore(null, "8.11.0"));
-
-    // then
-    assertThat(status.state()).isEqualTo(MigrationState.MIGRATION_IN_PROGRESS);
-    assertThat(status.detail()).contains("fresh database");
-  }
-
-  @Test
-  void shouldReportUnknownForIncompatibleUpgradePath() {
-    // given - schema=8.9.0, app=8.11.0 (skipped 8.10) - a real problem, not "in progress"
-    final var status = singleStatus(versionStore("8.9.0", "8.11.0"));
-
-    // then
-    assertThat(status.state()).isEqualTo(MigrationState.UNKNOWN);
-  }
-
-  @Test
-  void shouldReportUnknownForIndeterminateSchemaVersion() {
-    // given - stored schema version is not a valid semantic version
-    final var status = singleStatus(versionStore("not-a-semver", "8.10.0"));
-
-    // then
-    assertThat(status.state()).isEqualTo(MigrationState.UNKNOWN);
-  }
-
-  @Test
-  void shouldReportUnknownForUnparseableApplicationVersion() {
-    // given - app=development (not a semantic version)
-    final var status = singleStatus(versionStore("8.9.0", "development"));
-
-    // then
-    assertThat(status.state()).isEqualTo(MigrationState.UNKNOWN);
-    assertThat(status.detail()).contains("development");
-  }
-
-  @Test
-  void shouldReportUnknownWhenReadingCurrentVersionFails() {
-    // given - the search engine call itself fails (e.g. connection refused)
-    final var searchEngineClient = mock(SearchEngineClient.class);
-    when(searchEngineClient.indexExists(anyString()))
-        .thenThrow(new RuntimeException("connection refused"));
-    final var store = new ElasticsearchSchemaVersionStore(searchEngineClient, "", true, "8.10.0");
-
-    // when
-    final var status = singleStatus(store);
-
-    // then - a read failure must never throw; it must be reported as UNKNOWN
-    assertThat(status.state()).isEqualTo(MigrationState.UNKNOWN);
-    assertThat(status.detail()).contains("connection refused");
   }
 
   private static ElasticsearchSchemaMigrationStatusProvider provider(
