+++ schema-manager/src/test/java/io/camunda/search/schema/ElasticsearchSchemaVersionStoreTest.java
@@ -7,13 +7,13 @@
  */
 package io.camunda.search.schema;
 
-import static io.camunda.search.schema.ElasticsearchSchemaVersionStore.CurrentSchemaVersion.Kind;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
 import io.camunda.webapps.schema.descriptors.index.MetadataIndex;
+import io.camunda.zeebe.util.migration.CurrentSchemaVersion.Kind;
 import java.util.Map;
 import org.junit.jupiter.api.Test;
